requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langchain-google-genai==1.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
ollama==0.1.7
openai==1.6.1
gunicorn==21.2.0